requirements.txt
bs4
requests
PyPDF2
langchain
sentence-transformers
faiss-cpu
google-generativeai
fastapi
uvicorn
python-dotenv
pydantic
ultralytics
opencv-python
cv2
faster-whisper 
ffmpeg-python


